+++ test/contrib/test_socks.py
@@ -7,7 +7,7 @@ from urllib3.exceptions import ConnectTimeoutError, NewConnectionError
 from dummyserver.server import DEFAULT_CERTS
 from dummyserver.testcase import IPV4SocketDummyServerTestCase
 
-from nose.plugins.skip import SkipTest
+import pytest
 
 try:
     import ssl
@@ -194,6 +194,14 @@ def handle_socks4_negotiation(sock, username=None):
     yield True  # Avoid StopIteration exceptions getting fired.
 
 
+class TestSOCKSProxyManager(object):
+
+    def test_invalid_socks_version_is_valueerror(self):
+        with pytest.raises(ValueError) as e:
+            socks.SOCKSProxyManager(proxy_url='http://example.org')
+        assert 'Unable to determine SOCKS version' in e.value.args[0]
+
+
 class TestSocks5Proxy(IPV4SocketDummyServerTestCase):
     """
     Test the SOCKS proxy in SOCKS5 mode.
@@ -206,7 +214,7 @@ class TestSocks5Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, '16.17.18.19')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             while True:
@@ -238,7 +246,7 @@ class TestSocks5Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertIn(addr, ['127.0.0.1', '::1'])
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             while True:
@@ -270,7 +278,7 @@ class TestSocks5Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, b'example.com')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             buf = b''
@@ -364,7 +372,7 @@ class TestSocks5Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, '16.17.18.19')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             while True:
@@ -382,6 +390,45 @@ class TestSocks5Proxy(IPV4SocketDummyServerTestCase):
         proxy_url = "socks5://%s:%s" % (self.host, self.port)
         pm = socks.SOCKSProxyManager(proxy_url, username='user',
                                      password='pass')
+        self.addCleanup(pm.clear)
+
+        response = pm.request('GET', 'http://16.17.18.19')
+
+        self.assertEqual(response.status, 200)
+        self.assertEqual(response.data, b'')
+        self.assertEqual(response.headers['Server'], 'SocksTestServer')
+
+    def test_socks_with_auth_in_url(self):
+        """
+        Test when we have auth info in url, i.e.
+        socks5://user:pass@host:port and no username/password as params
+        """
+        def request_handler(listener):
+            sock = listener.accept()[0]
+
+            handler = handle_socks5_negotiation(
+                sock, negotiate=True, username=b'user', password=b'pass'
+            )
+            addr, port = next(handler)
+
+            self.assertEqual(addr, '16.17.18.19')
+            self.assertEqual(port, 80)
+            handler.send(True)
+
+            while True:
+                buf = sock.recv(65535)
+                if buf.endswith(b'\r\n\r\n'):
+                    break
+
+            sock.sendall(b'HTTP/1.1 200 OK\r\n'
+                         b'Server: SocksTestServer\r\n'
+                         b'Content-Length: 0\r\n'
+                         b'\r\n')
+            sock.close()
+
+        self._start_server(request_handler)
+        proxy_url = "socks5://user:pass@%s:%s" % (self.host, self.port)
+        pm = socks.SOCKSProxyManager(proxy_url)
         self.addCleanup(pm.clear)
 
         response = pm.request('GET', 'http://16.17.18.19')
@@ -424,7 +471,7 @@ class TestSocks5Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, '16.17.18.19')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             while True:
@@ -463,7 +510,7 @@ class TestSOCKS4Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, '16.17.18.19')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             while True:
@@ -495,7 +542,7 @@ class TestSOCKS4Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, '127.0.0.1')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             while True:
@@ -527,7 +574,7 @@ class TestSOCKS4Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, b'example.com')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             buf = b''
@@ -584,7 +631,7 @@ class TestSOCKS4Proxy(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, '16.17.18.19')
-            self.assertTrue(port, 80)
+            self.assertEqual(port, 80)
             handler.send(True)
 
             while True:
@@ -632,10 +679,8 @@ class TestSOCKSWithTLS(IPV4SocketDummyServerTestCase):
     """
     Test that TLS behaves properly for SOCKS proxies.
     """
+    @pytest.mark.skipif(not HAS_SSL, reason='No TLS available')
     def test_basic_request(self):
-        if not HAS_SSL:
-            raise SkipTest("No TLS available")
-
         def request_handler(listener):
             sock = listener.accept()[0]
 
@@ -643,7 +688,7 @@ class TestSOCKSWithTLS(IPV4SocketDummyServerTestCase):
             addr, port = next(handler)
 
             self.assertEqual(addr, b'localhost')
-            self.assertTrue(port, 443)
+            self.assertEqual(port, 443)
             handler.send(True)
 
             # Wrap in TLS