+++ test/test_util.py
@@ -11,6 +11,7 @@ import pytest
 
 from urllib3 import add_stderr_logger, disable_warnings
 from urllib3.util.request import make_headers, rewind_body, _FAILEDTELL
+from urllib3.util.response import assert_header_parsing
 from urllib3.util.retry import Retry
 from urllib3.util.timeout import Timeout
 from urllib3.util.url import (
@@ -499,7 +500,8 @@ class TestUtil(object):
         ssl_.HAS_SNI = False
         try:
             with patch('warnings.warn') as warn:
-                ssl_wrap_socket(ssl_context=mock_context, sock=socket)
+                ssl_wrap_socket(ssl_context=mock_context, sock=socket,
+                                server_hostname='www.google.com')
             mock_context.wrap_socket.assert_called_once_with(socket)
             assert warn.call_count >= 1
             warnings = [call[0][1] for call in warn.call_args_list]
@@ -565,3 +567,12 @@ class TestUtil(object):
     def test_parse_retry_after(self, value, expected):
         retry = Retry()
         assert retry.parse_retry_after(value) == expected
+
+    @pytest.mark.parametrize('headers', [
+        b'foo',
+        None,
+        object,
+    ])
+    def test_assert_header_parsing_throws_typeerror_with_non_headers(self, headers):
+        with pytest.raises(TypeError):
+            assert_header_parsing(headers)